:technologist: Implement test_step to generate submission CSV file (#9)
* :heavy_plus_sign: Add trimesh

Python library for loading and using triangular meshes!

* :technologist: Implement test_step to generate submission CSV file

Produce CSV files for submission to HuggingFace! Specifically, convert the predicted masks on the validation set to binary run length encoding format, and format it into a dataframe table. Documented how to create the output CSV file with a one-liner `python trainer.py test` command, and added unit tests to ensure that the implementation in ChaBuDNet.test_step works. Also gitignoring CSV files, and modified buffer of the demux DataPipe from 1024 to 2048.